+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState.hs
@@ -154,17 +154,4 @@ import Cardano.Ledger.Shelley.PParams (pvCanFollow)
 import Cardano.Ledger.Shelley.RewardUpdate
 import Cardano.Ledger.Shelley.Rules.Ppup (ShelleyGovState (..))
 import Cardano.Ledger.Shelley.UTxO (consumed, produced)
-import Cardano.Ledger.State (
-  CanGetUTxO (..),
-  CanSetUTxO (..),
-  UTxO (..),
-  ssDelegationsL,
-  ssFeeL,
-  ssPoolParamsL,
-  ssStakeDistrL,
-  ssStakeGoL,
-  ssStakeL,
-  ssStakeMarkL,
-  ssStakeMarkPoolDistrL,
-  ssStakeSetL,
- )
+import Cardano.Ledger.State