+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/TreeDiff.hs
@@ -126,7 +126,6 @@ instance ToExpr PulsingRewUpdate where
 instance ToExpr RewardUpdate
 
 -- LedgerState/Types
-instance ToExpr AccountState
 
 instance
   ( ToExpr (TxOut era)